Update report.md
The following has been removed from the report as the report was not completed:

## Progress

This is intended WBS and schedule. This can change as we proceed as the
proposed plan below and will evolve.

1. Week of 28th Oct: Architecture, Design finalization.
1. Week of 4th Nov:

* API signature definition and scope freeze (This should define how
  many APIs will be)
* Environment creation and Code Framework Creation (Skeleton, stub
  and dummy test cases)

1. Week of 11th Nov:  Selected APIs for VM
1. Week of 18th Nov:  Selected APIs for Image
1. Week of 25th Nov:  Selected APIs for Security
1. Week of 2nd Dec:

   * Buffer
   * Documentation completion
   * Pending PyTest (if any) completion

## Results

## Benchmark

TBD Benchmarking will be added as the interfaces are defined

## Testing

TBD Pytests will be added to the interfaces, as the APIs are developed.